requirements.txt
pylint==3.0.2
Flask==3.0.3
requests==2.32.2
pytest==8.2.1
python-dotenv==1.0.1
prometheus_client==0.20.0
pymemcache==4.0.0